move static data to folder
