+++ server/data/static/line-routes/dual-path-line-route.ts
