+++ server/data/static/line-routes/simple-line-route.ts
