+++ server/data/static/station.ts
